nov2022/src/test/java/com/n2/digitalwallet2/AccountTest.java
package com.n2.digitalwallet2;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

/**
 * One potential improvement would be to add some error handling in case any thread encounters an exception while executing its task. For example, you could catch any RuntimeExceptions and log them to the console or a log file.
 *
 * Additionally, it might be useful to add some timing information to measure the performance of the system, such as how long it takes to complete all the transactions or how long each individual transaction takes to complete. This information could be used to optimize the system for better performance in the future.
 *
 * When testing concurrent programs, it's important to consider the potential race conditions and thread synchronization issues that can arise. Here are some good practices for testing concurrent programs:
 *
 * Write deterministic tests: The outcome of the tests should be deterministic, meaning that they should produce the same result every time they are run.
 *
 * Use appropriate tools: There are several tools and libraries available for testing concurrent programs, such as JUnit, TestNG, and Thread Weaver. These tools can help you write and run tests that simulate different types of concurrent scenarios.
 *
 * Test with different thread counts: To ensure that the program behaves correctly under different loads, it's important to test it with different thread counts. This will help you identify any issues related to thread synchronization and race conditions.
 *
 * Test with different input values: To ensure that the program behaves correctly with different input values, it's important to test it with a variety of input values. This will help you identify any issues related to data consistency and thread synchronization.
 *
 * Use synchronization primitives: When testing concurrent programs, it's important to use synchronization primitives such as locks, semaphores, and barriers to ensure that threads are properly synchronized and that data consistency is maintained.
 *
 * Use mock objects: To isolate the program under test from its dependencies, it's important to use mock objects for testing. This will help you simulate different scenarios and ensure that the program behaves correctly under different conditions.
 *
 * Test for performance: Finally, it's important to test the program for performance to ensure that it can handle the expected load. This can be done by measuring response times and throughput under different loads and input values.
 *
 * By following these good practices, you can ensure that your unit and integration tests for concurrent programs are effective and thorough.
 */
public class AccountTest {

  @Mock
  DigitalWallet digitalWallet;

  @BeforeEach
  void setup() {
    MockitoAnnotations.openMocks(this);
  }

  @Test
  void testDeposit() {
    Account account = new Account(1, 10);
    account.deposit(100);
    assertEquals(110, account.getBalance());
  }

  @Test
  void testWithdraw() {
    Account account = new Account(1, 100);
    boolean result = account.withdraw(50);
    assertTrue(result);
    assertEquals(50, account.getBalance());
  }

  @Test
  void testTransfer() {
    Account fromAccount = new Account(1, 100);
    Account toAccount = new Account(2, 10);
    boolean result = fromAccount.transfer(toAccount, 50);
    assertTrue(result);
    assertEquals(50, fromAccount.getBalance());
    assertEquals(60, toAccount.getBalance());
  }
}